wardrobe_service/requirements.txt
# ── Core ML / PyTorch stack ─────────────────────────────────────────
torch==2.0.1
torchvision==0.15.2
timm==0.9.12

# ── TensorFlow / Keras stack (Wardrobe AI) ──────────────────────────
tensorflow==2.19.0
keras==3.10.0

# ── Image Processing ────────────────────────────────────────────────
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78   # optional fallback for server environments
albumentations==1.3.1
Pillow==10.1.0
rembg==2.0.50
onnxruntime==1.16.3

# ── Web Framework ───────────────────────────────────────────────────
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# ── Data Processing & ML utilities ──────────────────────────────────
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.5.0
h5py==3.11.0

# ── Shared utilities ────────────────────────────────────────────────
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0
huggingface-hub==0.19.4